refactor:앨범 생성 기능을 서비스, 훅으로 분리
