+++ src/pages/AlbumEditor.jsx
@@ -1,9 +1,7 @@
 import AlbumTitleForm from '@/components/AlbumEditor/AlbumTitleForm'
 import Grid from '@/components/common/Grid'
-import axios from 'axios' // 추가: axios import
 import { useCallback, useMemo, useState } from 'react'
-import { useNavigate, useParams } from 'react-router-dom'
-import { v4 as uuidv4 } from 'uuid'
+import { useParams } from 'react-router-dom'
 import CreateAlbumButton from '../components/AlbumEditor/CreateAlbumButton'
 
 // 커스텀 컴포넌트와 훅
@@ -12,11 +10,11 @@ import useFileUpload from '../hooks/useFileUpload'
 import { validateImageFiles } from '../services/validateImageFile'
 
 // APIs
-import { createAlbum, getPreSignedUrl } from '../api/album'
 
 // Assets
 import crossIcon from '../assets/cross_icon.png'
 import AlbumEditorHeader from '../components/AlbumEditor/AlbumEditorHeader'
+import { useAlbumCreation } from '../hooks/useAlbumCreation'
 import { useAlbumTitle } from '../hooks/useAlbumTitle'
 
 // 파일 미리보기 컴포넌트
@@ -62,13 +60,11 @@ const Alert = ({ message, type = 'error', onAction, actionText }) => {
 
 const AlbumEditor = () => {
     const { albumId } = useParams()
-    const navigate = useNavigate()
 
-    const [loading, setLoading] = useState(false)
     const [customError, setCustomError] = useState(null)
 
     const { albumTitle, handleTitleChange } = useAlbumTitle()
-
+    const { loading, error: albumError, createAlbumWithFiles, setError: setAlbumError } = useAlbumCreation()
     const {
         files,
         addFile,
@@ -108,100 +104,108 @@ const AlbumEditor = () => {
         },
         [addFile]
     )
-    const handleCreateAlbum = async () => {
+    // const handleCreateAlbum = async () => {
+    //     if (files.length === 0 || loading || isProcessing) return
+
+    //     const trimmedTitle = albumTitle.trim()
+    //     if (!trimmedTitle) {
+    //         setCustomError('앨범 제목을 입력해주세요.')
+    //         return
+    //     }
+
+    //     setLoading(true)
+    //     setCustomError(null)
+
+    //     try {
+    //         // 1. 앨범 이름과 파일 메타데이터 준비 - 각 파일에 새 이름 할당
+    //         const filesWithNewNames = files.map(fileItem => {
+    //             const newName = uuidv4() + '.' + fileItem.file.type.split('/')[1]
+    //             return {
+    //                 ...fileItem,
+    //                 newName,
+    //                 originalFile: fileItem.file,
+    //             }
+    //         })
+
+    //         const pictures = filesWithNewNames.map(fileItem => ({
+    //             pictureName: fileItem.newName,
+    //             pictureType: fileItem.originalFile.type,
+    //         }))
+
+    //         console.log('생성된 pictures:', pictures)
+
+    //         // 2. presigned URL 요청
+    //         const response = await getPreSignedUrl({ pictures })
+    //         const presignedFiles = response.data.presignedFiles
+
+    //         console.log('서버에서 받은 presignedFiles:', presignedFiles)
+
+    //         // 3. 각 파일을 presigned URL을 사용하여 업로드
+    //         for (const fileItem of filesWithNewNames) {
+    //             const file = fileItem.originalFile
+    //             const newName = fileItem.newName
+
+    //             const matched = presignedFiles.find(f => f.pictureName === newName)
+
+    //             if (!matched) {
+    //                 console.error(`${newName}에 대한 매칭되는 presigned URL을 찾을 수 없습니다.`)
+    //                 continue
+    //             }
+
+    //             console.log(`${newName} 파일 업로드 시작...`)
+
+    //             try {
+    //                 await axios.put(matched.presignedUrl, file, {
+    //                     headers: {
+    //                         'Content-Type': file.type,
+    //                     },
+    //                 })
+    //                 console.log(`${newName} 파일 업로드 완료!`)
+    //             } catch (error) {
+    //                 console.error(`${newName} 파일 업로드 중 오류:`, error)
+    //                 throw error // 오류를 상위로 전파
+    //             }
+    //         }
+
+    //         // 4. 업로드 완료 후 앨범 생성 요청
+    //         // pictureUrl은 S3의 URL로 수정해야 함, presignedUrl은 업로드용이지 액세스용이 아님
+
+    //         const pictureData = presignedFiles.map(f => ({
+    //             pictureUrl: f.pictureURL || f.pictureName, // 서버 응답에 따라 적절한 필드 사용
+    //             latitude: 0.0,
+    //             longitude: 0.0,
+    //         }))
+
+    //         const albumData = {
+    //             albumName: trimmedTitle,
+    //             pictureUrls: pictureData,
+    //         }
+
+    //         console.log('생성할 앨범 데이터:', albumData)
+    //         if (!albumId) {
+    //             const result = await createAlbum(albumData)
+    //             console.log('앨범 생성 결과:', result)
+    //             navigate('/main')
+    //         } else {
+    //             const result = await addAlbumPicture(albumId, albumData)
+    //             console.log('사진 추가 결과:', result)
+    //             navigate(`/album/${albumId}`)
+    //         }
+    //     } catch (err) {
+    //         console.error('전체 오류:', err)
+    //         setCustomError('앨범 생성 중 오류가 발생했습니다.')
+    //     } finally {
+    //         setLoading(false)
+    //     }
+    // }
+
+    // 앨범 생성 핸들러 - 이제 단순히 훅 호출만
+    const handleCreateAlbum = useCallback(async () => {
         if (files.length === 0 || loading || isProcessing) return
 
-        const trimmedTitle = albumTitle.trim()
-        if (!trimmedTitle) {
-            setCustomError('앨범 제목을 입력해주세요.')
-            return
-        }
-
-        setLoading(true)
-        setCustomError(null)
-
-        try {
-            // 1. 앨범 이름과 파일 메타데이터 준비 - 각 파일에 새 이름 할당
-            const filesWithNewNames = files.map(fileItem => {
-                const newName = uuidv4() + '.' + fileItem.file.type.split('/')[1]
-                return {
-                    ...fileItem,
-                    newName,
-                    originalFile: fileItem.file,
-                }
-            })
-
-            const pictures = filesWithNewNames.map(fileItem => ({
-                pictureName: fileItem.newName,
-                pictureType: fileItem.originalFile.type,
-            }))
-
-            console.log('생성된 pictures:', pictures)
-
-            // 2. presigned URL 요청
-            const response = await getPreSignedUrl({ pictures })
-            const presignedFiles = response.data.presignedFiles
-
-            console.log('서버에서 받은 presignedFiles:', presignedFiles)
-
-            // 3. 각 파일을 presigned URL을 사용하여 업로드
-            for (const fileItem of filesWithNewNames) {
-                const file = fileItem.originalFile
-                const newName = fileItem.newName
-
-                const matched = presignedFiles.find(f => f.pictureName === newName)
-
-                if (!matched) {
-                    console.error(`${newName}에 대한 매칭되는 presigned URL을 찾을 수 없습니다.`)
-                    continue
-                }
-
-                console.log(`${newName} 파일 업로드 시작...`)
-
-                try {
-                    await axios.put(matched.presignedUrl, file, {
-                        headers: {
-                            'Content-Type': file.type,
-                        },
-                    })
-                    console.log(`${newName} 파일 업로드 완료!`)
-                } catch (error) {
-                    console.error(`${newName} 파일 업로드 중 오류:`, error)
-                    throw error // 오류를 상위로 전파
-                }
-            }
-
-            // 4. 업로드 완료 후 앨범 생성 요청
-            // pictureUrl은 S3의 URL로 수정해야 함, presignedUrl은 업로드용이지 액세스용이 아님
-
-            const pictureData = presignedFiles.map(f => ({
-                pictureUrl: f.pictureURL || f.pictureName, // 서버 응답에 따라 적절한 필드 사용
-                latitude: 0.0,
-                longitude: 0.0,
-            }))
-
-            const albumData = {
-                albumName: trimmedTitle,
-                pictureUrls: pictureData,
-            }
-
-            console.log('생성할 앨범 데이터:', albumData)
-            if (!albumId) {
-                const result = await createAlbum(albumData)
-                console.log('앨범 생성 결과:', result)
-                navigate('/main')
-            } else {
-                const result = await addAlbumPicture(albumId, albumData)
-                console.log('사진 추가 결과:', result)
-                navigate(`/album/${albumId}`)
-            }
-        } catch (err) {
-            console.error('전체 오류:', err)
-            setCustomError('앨범 생성 중 오류가 발생했습니다.')
-        } finally {
-            setLoading(false)
-        }
-    }
+        // 앨범 생성 로직을 훅에 위임
+        await createAlbumWithFiles(albumTitle, files, albumId)
+    }, [albumTitle, files, albumId, createAlbumWithFiles, loading, isProcessing])
 
     // 초과 파일 추가 핸들러
     const handleAddOverflowFiles = useCallback(() => {
